pyproject.toml
[project]
name = "agent-ai-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.4"
dependencies = [
    "anthropic>=0.68.0",
    "beautifulsoup4>=4.13.5",
    "gradio>=5.49.1",
    "isort>=6.0.1",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.1",
]
